Fix support for leading zeros in source numbers
As per the example, source numbers should allow for leading zeros both
in the spectrum file name as well as in the spectrum overrides. The dot
separated source identifier in the spectrum overrides is parsed as a
string, which is inflexible with respect to leading zeros.

Parse the source number as an integer in the spectrum overrides.